loc(12:47)7/8/2023

+++ app/src/main/java/com/example/duan1bookapp/fragment/Fragment_Home.java
@@ -254,7 +254,7 @@ private void loadTrendingBooks() {
                     for (DataSnapshot ds : snapshot.getChildren()) {
                         //get data
                         ModelPdfTrendingBooks model = ds.getValue(ModelPdfTrendingBooks.class);
-                        int viewCount= (int) model.getViewsCount();
+                        long viewCount=model.getViewsCount();
                         if(viewCount >= 2){
                             //add to list
                             pdfTrendingBooksList.add(model);
